support for extended channels by using frequencies instead of channels. this change requires also a wireless extension change to support private calls with bigger data arrays
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@6212 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/services/wireless_copy.h
@@ -1065,8 +1065,8 @@ struct	iw_priv_args
  */
 struct iw_event
 {
-	__u16		len;			/* Real lenght of this stuff */
-	__u16		cmd;			/* Wireless IOCTL */
+	__u32		len;			/* Real lenght of this stuff */
+	__u32		cmd;			/* Wireless IOCTL */
 	union iwreq_data	u;		/* IOCTL fixed payload */
 };
